requirements.txt
Django
South
django-annoying
MySQL-python
PIL
django-grappelli